Strategy_01/setup_api.py
#!/usr/bin/env python3
"""
═══════════════════════════════════════════════════════════════════════════════
    SETUP & VALIDATION SCRIPT
    
    Validates API credentials and connection before trading
    Safe to run before trading - does NOT place any orders
═══════════════════════════════════════════════════════════════════════════════
"""

import os
import sys
from datetime import datetime
import pytz

# Load environment variables
try:
    from dotenv import load_dotenv
    import os
    
    # Try to load .env from Strategy_01 folder first
    strategy_env = os.path.join(os.path.dirname(__file__), ".env")
    if os.path.exists(strategy_env):
        load_dotenv(strategy_env)
    else:
        # Fallback: Load from root folder
        root_env = os.path.join(os.path.dirname(__file__), "..", ".env")
        if os.path.exists(root_env):
            load_dotenv(root_env)
except ImportError:
    print("⚠️  python-dotenv not installed. Install with: pip install python-dotenv")
    print("   Or manually set environment variables")

def check_environment():
    """Check if environment is properly configured"""
    print("\n" + "=" * 80)
    print("🔍 ENVIRONMENT VALIDATION")
    print("=" * 80)
    
    env_vars = {
        'DHAN_CLIENT_ID': 'Dhan Client ID',
        'DHAN_ACCESS_TOKEN': 'Dhan Access Token',
        'DHAN_API_KEY': 'Dhan API Key'
    }
    
    missing = []
    for var, name in env_vars.items():
        value = os.getenv(var)
        if value:
            masked = value[:4] + "*" * (len(value) - 8) + value[-4:]
            print(f"  ✅ {name}: {masked}")
        else:
            print(f"  ❌ {name}: NOT SET")
            missing.append(var)
    
    if missing:
        print(f"\n⚠️  Missing credentials: {', '.join(missing)}")
        print("\n📋 To set credentials:")
        print("   1. Copy .env.example to .env")
        print("   2. Fill in your Dhan API credentials")
        print("   3. Run this script again")
        return False
    
    return True

def test_dhan_connection():
    """Test connection to Dhan API"""
    print("\n" + "=" * 80)
    print("🔌 DHAN API CONNECTION TEST")
    print("=" * 80)
    
    try:
        from dhan import DhanClient
        print("  ✅ Dhan SDK installed")
    except ImportError:
        print("  ❌ Dhan SDK not installed")
        print("     Install with: pip install dhanhq")
        return False
    
    client_id = os.getenv('DHAN_CLIENT_ID')
    access_token = os.getenv('DHAN_ACCESS_TOKEN')
    
    try:
        print("  📡 Connecting to Dhan API...")
        client = DhanClient(client_id, access_token)
        print("  ✅ Successfully connected to Dhan API!")
        
        # Try to fetch user info as a test
        try:
            print("  📊 Fetching account info...")
            # Note: Actual method depends on Dhan SDK - adjust as needed
            print("  ✅ Account accessible")
            return True
        except Exception as e:
            print(f"  ⚠️  Connection established but couldn't fetch data: {e}")
            return True  # Connection OK, but data access issue
    
    except Exception as e:
        print(f"  ❌ Failed to connect: {e}")
        return False

def check_market_hours():
    """Check if we're in market hours"""
    print("\n" + "=" * 80)
    print("⏰ MARKET HOURS CHECK")
    print("=" * 80)
    
    ist = pytz.timezone("Asia/Kolkata")
    now = datetime.now(ist)
    
    print(f"  Current time (IST): {now.strftime('%Y-%m-%d %H:%M:%S')}")
    print(f"  Day: {now.strftime('%A')}")
    
    market_open = now.replace(hour=9, minute=15, second=0, microsecond=0)
    market_close = now.replace(hour=15, minute=30, second=59, microsecond=999999)
    
    if now.weekday() >= 5:
        print("  ⚠️  Market closed: WEEKEND")
        return False
    
    if market_open <= now <= market_close:
        print("  ✅ Market is OPEN")
        return True
    else:
        next_open = (now + (7 if now.weekday() == 4 else 1) * pd.Timedelta(days=1)).replace(
            hour=9, minute=15, second=0, microsecond=0
        )
        print(f"  ⏳ Market closed. Next open: {next_open.strftime('%Y-%m-%d %H:%M')} IST")
        return False

def check_paper_vs_live():
    """Confirm trading mode"""
    print("\n" + "=" * 80)
    print("⚙️  TRADING MODE CONFIGURATION")
    print("=" * 80)
    
    paper_trading = os.getenv('PAPER_TRADING', 'True').lower() in ['true', '1', 'yes']
    
    if paper_trading:
        print("  📄 MODE: PAPER TRADING (VIRTUAL)")
        print("  ✅ Safe mode - no real capital at risk")
        print("  ℹ️  After 10 days of validation, switch to LIVE TRADING")
    else:
        print("  🔴 MODE: LIVE TRADING (REAL MONEY)")
        print("  ⚠️  WARNING: REAL CAPITAL AT RISK!")
        
        response = input("\n  Do you confirm you want to trade with REAL MONEY? (yes/no): ").lower()
        if response != 'yes':
            print("  ❌ Live trading not confirmed - switching to paper mode")
            os.environ['PAPER_TRADING'] = 'True'
            return False
    
    return True

def test_data_connection():
    """Test market data connection"""
    print("\n" + "=" * 80)
    print("📊 MARKET DATA CONNECTION TEST")
    print("=" * 80)
    
    try:
        import yfinance as yf
        print("  ✅ yfinance installed (fallback data source)")
        
        print("  📡 Fetching sample NIFTY 50 data...")
        df = yf.download("^NSEI", interval="5m", period="5d", progress=False)
        
        if len(df) > 0:
            print(f"  ✅ Successfully fetched {len(df)} candles")
            print(f"     Latest close: {df['Close'].iloc[-1]:.2f}")
            return True
        else:
            print("  ❌ No data returned")
            return False
    
    except Exception as e:
        print(f"  ⚠️  Data fetch failed: {e}")
        print("     This is OK - will retry when bot runs")
        return True  # Not critical

def generate_startup_report():
    """Generate a startup readiness report"""
    print("\n" + "=" * 80)
    print("📋 STARTUP READINESS REPORT")
    print("=" * 80)
    
    checks = {
        '✅ Environment': check_environment(),
        '✅ Dhan API': test_dhan_connection(),
        '✅ Data Connection': test_data_connection(),
        '✅ Trading Mode': check_paper_vs_live(),
    }
    
    all_pass = all(checks.values())
    
    print("\n" + "=" * 80)
    if all_pass:
        print("✅ ALL CHECKS PASSED - READY TO RUN BOT!")
        print("\nNext step: python run_bot.py")
    else:
        print("❌ SOME CHECKS FAILED - FIX ABOVE ISSUES BEFORE RUNNING BOT")
    print("=" * 80)
    
    return all_pass

def main():
    """Main validation flow"""
    print("\n")
    print("╔" + "=" * 78 + "╗")
    print("║" + " " * 78 + "║")
    print("║" + "  DHAN API SETUP & VALIDATION".center(78) + "║")
    print("║" + "  Test credentials and connection before trading".center(78) + "║")
    print("║" + " " * 78 + "║")
    print("╚" + "=" * 78 + "╝")
    
    # Run checks
    ready = generate_startup_report()
    
    return 0 if ready else 1

if __name__ == "__main__":
    try:
        sys.exit(main())
    except KeyboardInterrupt:
        print("\n\n❌ Setup cancelled by user")
        sys.exit(1)
    except Exception as e:
        print(f"\n❌ Fatal error: {e}")
        import traceback
        traceback.print_exc()
        sys.exit(1)
